back/requirements.txt
asgiref==3.3.4
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
fastapi==0.65.1
h11==0.12.0
httptools==0.2.0
mongoengine==0.23.1
pycparser==2.20
pydantic==1.8.2
pymongo==3.11.4
python-dotenv==0.17.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
